+++ data/scripts/sector/managestationincomes.lua
@@ -144,7 +144,7 @@ function ManageStationIncomes.giveStationMoney(station, _seller)
     if not faction then return end
 
     -- Cosmic Overhaul Balance tweak: Reduced base payout from 16k to 8k
-    local money = math.floor((0.3+(2*random():getInt()/3))*8000)
+    local money = math.floor((0.3+(2*random():getFloat()/3))*8000)
     if random():getFloat(0, 1) < 0.2 then money = money*2 end
     money = math.floor(money*mapping.quantity)
     money = money*ManageStationIncomes.getWarHeatMultiplier()
@@ -187,7 +187,7 @@ function ManageStationIncomes.getResourceIncome()
         if probFactor > 0.05 then
             local matRichness = math.max(probFactor*(richness), 0.2)
             -- Cosmic Overhaul Balance tweak: Reduced from 8000 base to 3500 base
-            mats = (0.5+random():getInt()/2)*3500
+            mats = (0.5+random():getFloat()/2)*3500
             mats = mats*matRichness
             mats = mats*ManageStationIncomes.getWarHeatMultiplier()
             mats = mats*payoutMult
